Refactor, combine types into types.rs, combine logic into lib.rs
