+++ src/main.rs
@@ -1,8 +1,8 @@
 extern crate swapi;
 
 fn main() {
-    let mut planet_resp: swapi::types::species::Species = Default::default();
+    let mut planet_resp: swapi::types::Species = Default::default();
     println!("{:#?}", planet_resp);
-    swapi::types::species::query_species("6", &mut planet_resp);
+    swapi::query_species("6", &mut planet_resp);
     println!("{:#?}", planet_resp);
 }
