--- src/types/mod.rs
@@ -1,7 +1,0 @@
-pub mod films;
-pub mod people;
-pub mod planets;
-pub mod species;
-pub mod starships;
-pub mod vehicles;
-use super::query;
